requirements.txt
requests==2.25.1
pandas==1.2.4
beautifulsoup4==4.9.3
datetime==4.3
pymongo==3.5.1
dnspython==2.1.0